+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Create `HTMLCanvasElement` when parsing `<canvas>` tags.
 * Create `HTMLDialogElement` when parsing `<dialog>` tags.
 * Fix parsing of `<ruby>` tags, especially `<rb>` and `<rtc>`.
+* Create `HTMLMenuItemElement` when parsing `<menuitem>` tags.
 
 # domino 1.0.23 (30 Jan 2016)
 * Fix `CSSStyleDeclaration#setProperty`. (#71)